feature(setup): host app on firebase
