+++ README.md
@@ -1,1 +1,3 @@
 # chat-backend
+
+Firebase App - https://chat-back-10336.firebaseapp.com/
