+++ firebase.json
@@ -1,0 +1,18 @@
+{
+  "hosting": {
+    "public": "dist",
+    "ignore": [
+      "firebase.json",
+      "**/.*",
+      "**/node_modules/**"
+    ],
+    "rewrites": [{
+      "source": "**",
+      "function": "app"
+    }]
+
+  },
+  "functions": {
+    "source": "functions"
+  }
+}
